usersmsa/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
django-mysql==4.10.0
djangorestframework==3.14.0
ez-setup==0.9
kafka-python==2.0.2
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
